cleanup, support printing floats & 64bit, clarify ISR code

+++ src/_test.cpp
@@ -6,8 +6,8 @@ long long addll(long long a, long long b) { return 62; }
 
 // call from C++ with parameter overloading
 int add(int a, int b) { return addi(a,b); }
-int add(long a, long b) { return addl(a,b); }
-int add(long long a, long long b) { return addll(a,b); }
+long add(long a, long b) { return addl(a,b); }
+long long add(long long a, long long b) { return addll(a,b); }
 
 void test() {
 	int c = 2;
